requirements.txt
asgiref==3.7.2
distro==1.8.0
Django==4.2.2
numpy==1.24.3
pandas==2.0.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tabula-py==2.7.0
typing_extensions==4.6.3
tzdata==2023.3
